Better error handling
It could be important to differentiate between a failed connection attempt
to a key server and the situation that a key was not found, but search()
returned None in both cases.

Therefore, this commit changes the behaviour from silently ignoring HTTP
connection errors to explicitly handling the case that a key was not found
(status code 404) in which case None is returned instead of a key or an error
occurred in which case an Exception is raised with detailed error information.